Updated Monthly Expense Report to have group By and subtotal feature

+++ expensetracker-reports/src/main/java/com/slabs/expense/tracker/reports/column/Column.java
@@ -1,12 +1,13 @@
 package com.slabs.expense.tracker.reports.column;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 public enum Column {
 
-	PRICEPERUNIT("Unit Price","PRICEPERUNIT", "pricePerUnit", Double.class), 
-	PRICE("Price", "PRICE", "price", Double.class), 
-	QTY("Qty","QTY", "qty", Double.class), 
+	PRICEPERUNIT("Unit Price","PRICEPERUNIT", "pricePerUnit",  BigDecimal.class), 
+	PRICE("Price", "PRICE", "price", BigDecimal.class), 
+	QTY("Qty","QTY", "qty",  BigDecimal.class), 
 	EXPTYPE("Expense Type", "EXPTYPE", "exptype", String.class), 
 	EXPDATE("Date", "EXPDATE", "expdate", Date.class),
 	ITEMNAME("Commodity", "ITEMNAME", "itemName", String.class),
@@ -15,7 +16,7 @@ public enum Column {
 	DESCRIPTION("Description", "DESCRIPTION", "description", String.class), 
 	CATEGORYID("Id", "CATEGORYID", "categoryId", Integer.class), 
 	ID("Id", "ID","id", Integer.class), 
-	INCOME("Income", "INCOME", "income", Double.class), 
+	INCOME("Income", "INCOME", "income",  BigDecimal.class), 
 	INCOMETYPE("Income Type", "INCOMETYPE", "incometype", String.class), 
 	YR("Year", "YR", "yr", Integer.class), 
 	MTH("Month", "MTH","mth", Integer.class),
